config: move yaml_dump into config
